add round rectangle from d2d and handle windows size and paint message

+++ source/draw/draw.h
@@ -4,10 +4,10 @@
 
 struct BaseRect
 {
-    long left;
-    long top;
-    long right;
-    long bottom;
+    float left;
+    float top;
+    float right;
+    float bottom;
 };
 
 struct NvpColor
